Final testing results
This spreadsheet contains the results of the three testing periods,
including the final test event.  All in all, the vehicle passed the
technical and operational testing requirements.  The addition of a text
display as a user interface was not accomplished so the requirements
related to that system were marked as no-test.  Good Job Team!  Very
well done!